Docker (#7)
* Fix compilation and execution paths to match

* Add a docker-compose setup

We add two dockerfiles (one for running the server and one for running
the client). The docker-compose file runs those two dockerfiles as
services as well as an unmodified mongodb service.

Some notes:

* create-react-app doesn't play nice with docker-compose. We had to add
a hack to the dockerfile to prevent it from continuously clearing the
scrollback--including logs from other containers
* We've placed appropriate environment variables on the containers,
though they are so far unused.
* Server and mongo are networked together, which will allow them to
communicate using a docker network (as far as the server is concerned,
the database's network URL is just `mongo`)

* Display a message on the client from the server

Done just as a proof of concept that server-client communication is 
possible. I have also added a .env.development file (which is 
automatically consumed by react-scripts in development) to ensure this 
will work even when run outside of Docker

See here for more details about react-scripts and .env files 
https://create-react-app.dev/docs/adding-custom-environment-variables/

* Ensure data fetching only occurs onxe

Passing [] to useEffect as dependencies indicates that data
fetching is safe to do once

* Reformat JSX conditional to be clearer

* Since ts enforces string value for serverMessage, it's safe just
  to check for truthiness
* Replace two conditions with a ternary